rename local/storage/ to local/tiers/ for clarity
Disambiguate from top-level storage/ (shared I/O infrastructure)
by renaming the standalone tier management layer to local/tiers/.

Made-with: Cursor
