+++ src/umbp/include/umbp/local/tiers/dram_tier.h
@@ -31,7 +31,7 @@
 #include <utility>
 #include <vector>
 
-#include "umbp/local/storage/tier_backend.h"
+#include "umbp/local/tiers/tier_backend.h"
 
 namespace mori::umbp {
