+++ src/umbp/include/umbp/local/tiers/spdk_proxy_tier.h
@@ -33,7 +33,7 @@
 #include <vector>
 
 #include "umbp/common/config.h"
-#include "umbp/local/storage/tier_backend.h"
+#include "umbp/local/tiers/tier_backend.h"
 #include "umbp/storage/spdk/proxy/spdk_proxy_protocol.h"
 #include "umbp/storage/spdk/proxy/spdk_proxy_shm.h"
