+++ src/umbp/include/umbp/local/tiers/ssd_tier.h
@@ -30,10 +30,10 @@
 #include <vector>
 
 #include "umbp/common/config.h"
-#include "umbp/local/storage/segment/segment_index.h"
-#include "umbp/local/storage/segment/segment_scanner.h"
-#include "umbp/local/storage/segment/segment_writer.h"
-#include "umbp/local/storage/tier_backend.h"
+#include "umbp/local/tiers/segment/segment_index.h"
+#include "umbp/local/tiers/segment/segment_scanner.h"
+#include "umbp/local/tiers/segment/segment_writer.h"
+#include "umbp/local/tiers/tier_backend.h"
 #include "umbp/storage/io/status.h"
 #include "umbp/storage/io/storage_io_driver.h"
