+++ src/umbp/local/tiers/dram_tier.cpp
@@ -19,7 +19,7 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
-#include "umbp/local/storage/dram_tier.h"
+#include "umbp/local/tiers/dram_tier.h"
 
 #include <fcntl.h>
 #include <sys/mman.h>
