+++ src/umbp/local/tiers/segment/segment_format.cpp
@@ -19,7 +19,7 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
-#include "umbp/local/storage/segment/segment_format.h"
+#include "umbp/local/tiers/segment/segment_format.h"
 
 #include <string>
