+++ src/umbp/local/tiers/segment/segment_writer.cpp
@@ -19,7 +19,7 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
-#include "umbp/local/storage/segment/segment_writer.h"
+#include "umbp/local/tiers/segment/segment_writer.h"
 
 namespace mori::umbp::segment {
