+++ src/umbp/local/tiers/tier_backend.cpp
@@ -19,7 +19,7 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
-#include "umbp/local/storage/tier_backend.h"
+#include "umbp/local/tiers/tier_backend.h"
 
 namespace mori::umbp {
