+++ tests/cpp/umbp/distributed/test_peer_service.cpp
@@ -34,7 +34,7 @@
 #include "umbp/distributed/peer/peer_service.h"
 #include "umbp/distributed/pool_client.h"
 #include "umbp/local/block_index/local_block_index.h"
-#include "umbp/local/storage/local_storage_manager.h"
+#include "umbp/local/tiers/local_storage_manager.h"
 #include "umbp_peer.grpc.pb.h"
 
 namespace mori::umbp {
