+++ tests/cpp/umbp/local/bench_spdk_raw.cpp
@@ -50,8 +50,8 @@
 
 #include "umbp/common/config.h"
 #include "umbp/common/log.h"
-#include "umbp/local/storage/spdk_ssd_tier.h"
-#include "umbp/local/storage/ssd_tier.h"
+#include "umbp/local/tiers/spdk_ssd_tier.h"
+#include "umbp/local/tiers/ssd_tier.h"
 #include "umbp/storage/spdk/spdk_env.h"
 
 using namespace mori::umbp;
