+++ tests/cpp/umbp/local/bench_umbp_micro.cpp
@@ -64,12 +64,12 @@
 
 #include "umbp/common/config.h"
 #include "umbp/local/standalone_client.h"
-#include "umbp/local/storage/dram_tier.h"
-#include "umbp/local/storage/local_storage_manager.h"
-#include "umbp/local/storage/segment/segment_format.h"
-#include "umbp/local/storage/spdk_proxy_tier.h"
-#include "umbp/local/storage/ssd_tier.h"
 #include "umbp/local/storage_tier.h"
+#include "umbp/local/tiers/dram_tier.h"
+#include "umbp/local/tiers/local_storage_manager.h"
+#include "umbp/local/tiers/segment/segment_format.h"
+#include "umbp/local/tiers/spdk_proxy_tier.h"
+#include "umbp/local/tiers/ssd_tier.h"
 #include "umbp/storage/io/storage_io_driver.h"
 
 using namespace mori::umbp;
