+++ tests/cpp/umbp/local/test_prefix_aware_eviction.cpp
@@ -27,7 +27,7 @@
 
 #include "umbp/local/block_index/local_block_index.h"
 #include "umbp/local/standalone_client.h"
-#include "umbp/local/storage/local_storage_manager.h"
+#include "umbp/local/tiers/local_storage_manager.h"
 
 using namespace mori::umbp;
