+++ tests/cpp/umbp/local/test_ssd_tier.cpp
@@ -24,7 +24,7 @@
 #include <iostream>
 #include <vector>
 
-#include "umbp/local/storage/local_storage_manager.h"
+#include "umbp/local/tiers/local_storage_manager.h"
 
 using namespace mori::umbp;
